chore: move `sign_transaction` example to `sign_transfer`
