+++ Cargo.toml
@@ -15,7 +15,8 @@ name = "get_version"
 name = "get_public_key"
 
 [[example]]
-name = "sign_transaction"
+name = "sign_transfer"
+path = "examples/sign_transaction/transfer.rs"
 
 [[example]]
 name = "blind_sign_transaction"
